requirements.txt
mlflow==1.9.1
PyYAML==5.1
scikit-learn==0.23.1
pandas==0.24.2
numpy==1.16.4